+++ test/background_migrations/copy_column_test.rb
@@ -5,6 +5,7 @@ require "test_helper"
 module BackgroundMigrations
   class CopyColumnTest < MiniTest::Test
     class Project < ActiveRecord::Base
+      default_scope { where(archived: false) }
     end
 
     def setup
@@ -12,13 +13,15 @@ module BackgroundMigrations
       @connection.create_table(:projects, id: :serial, force: :cascade) do |t|
         t.string :name
         t.text :settings
+        t.boolean :archived
         t.bigint :id_for_type_change
         t.jsonb :settings_for_type_change
       end
 
       Project.reset_column_information
 
-      @project = Project.create!(name: "rails", settings: '{"key":"value"}')
+      @project1 = Project.create!(name: "rails", settings: '{"key":"value"}')
+      @project2 = Project.create!(name: "postgresql")
     end
 
     def teardown
@@ -31,19 +34,22 @@ module BackgroundMigrations
     end
 
     def test_process_batch
-      m = OnlineMigrations::BackgroundMigrations::CopyColumn.new(:projects, ["id"], ["id_for_type_change"])
+      m = OnlineMigrations::BackgroundMigrations::CopyColumn.new(:projects, ["id"], ["id_for_type_change"], Project.name)
       m.process_batch(m.relation)
 
-      @project.reload
-      assert_equal @project.id, @project.id_for_type_change
+      @project1.reload
+      @project2.reload
+
+      assert_equal @project1.id, @project1.id_for_type_change
+      assert_equal @project2.id, @project2.id_for_type_change
     end
 
     def test_process_batch_type_cast_function
       m = OnlineMigrations::BackgroundMigrations::CopyColumn.new(:projects, ["settings"], ["settings_for_type_change"], nil, { "settings" => "jsonb" })
       m.process_batch(m.relation)
 
-      @project.reload
-      assert_equal "value", @project.settings_for_type_change["key"]
+      @project1.reload
+      assert_equal "value", @project1.settings_for_type_change["key"]
     end
 
     def test_count