+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Ignore default scopes in `CopyColumn` and `BackfillColumn` background migrations
 - Raise an error for unsupported database versions
 - Fix backfilling code in suggestion for changing column's NOT NULL